requirements.txt
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1
google-generativeai==0.7.2
fredapi==0.5.2
ta==0.11.0
schedule==1.2.2
rich==13.7.1